work on using labeledcheckbox and spinner at the same time
