+++ js/lib/multiply-widgets.js
@@ -1,5 +1,7 @@
 var checkbox = require('./checkbox.js')
+var spinner = require('./spinner.js')
 
 module.exports = {
+    SpinnerView : spinner.SpinnerView,
     LabeledCheckboxView : checkbox.LabeledCheckboxView
 };
