+++ js/lib/spinner.js
@@ -28,7 +28,3 @@ var SpinnerView = widgets.DOMWidgetView.extend({
     },
 
 });
-
-module.exports = {
-    SpinnerView : SpinnerView
-};
